+++ src/ast/passes/semantic_analyser.cpp
@@ -2841,7 +2841,7 @@ void SemanticAnalyser::visit(MapAccess &acc)
   } else {
     // If there is no record of any assignment after the first pass
     // then it's safe to say this map is undefined.
-    if (!is_first_pass()) {
+    if (!is_first_pass() && !acc.map->read_only) {
       acc.addError() << "Undefined map: " << acc.map->ident;
     }
     pass_tracker_.inc_num_unresolved();