+++ CHANGELOG.md
@@ -61,6 +61,8 @@ and this project adheres to
   - [#4254](https://github.com/bpftrace/bpftrace/pull/4254)
 - Add boolean values (`true` and `false`)
   - [#4280](https://github.com/bpftrace/bpftrace/pull/4280)
+- Add `getopt` function to handle named command line program arguments
+  - [#4258](https://github.com/bpftrace/bpftrace/pull/4258)
 #### Changed
 - `-p` CLI flag now applies to all probes (except BEGIN/END)
   - [#3800](https://github.com/bpftrace/bpftrace/pull/3800)